added help which opens instantly on docker run and changed that data can be downloaded in training script

+++ train.sh
@@ -3,8 +3,8 @@
 
 if [ $1 = "--help" ]
 then 
-  echo "Expected format: ./train modelname //expects data to be located under /data/models/modelname/wav"
-  echo "                 ./train modelname data-dir [utterance-postfix]"
+  echo "Expected format: ./train.sh modelname //expects data to be located under /data/models/modelname/wav"
+  echo "                 ./train.sh modelname data-dir [utterance-postfix]"
   echo "Available postfixes are:"
 echo "_Kinect-Beam
 _Samson
@@ -17,7 +17,7 @@ fi
 #TODO: check if data dir correct
 if [ $# -lt 1 ]
 then
-  echo "Not enough parameters. Expected format: ./train modelname data-dir [utterance-postfix]"
+  echo "Not enough parameters. Expected format: ./train.sh modelname data-dir [utterance-postfix]"
   echo "Available postfixes are:
 _Kinect-Beam
 _Samson
@@ -39,15 +39,23 @@ else
     #copy data to dir
     if [ $# -ge 2 ]
     then 
-      echo "Linking data from $2 to the data dir of the model"
+      if [ -f $2/download.sh ]
+      then
+        cd $2
+        echo "Downloading data"
+        ./download.sh
+      else
+
+        echo "Linking data from $2 to the data dir of the model"
       
-      ln -s $2/train /data/models/$1/s5/data/wav/
-      echo "Linked train"
-      ln -s $2/test /data/models/$1/s5/data/wav/
-      echo "Linked test"
-      ln -s $2/dev /data/models/$1/s5/data/wav/
-      echo "Linked dev"
-      cd /data/models/
+        ln -s $2/train /data/models/$1/s5/data/wav/
+        echo "Linked train"
+        ln -s $2/test /data/models/$1/s5/data/wav/
+        echo "Linked test"
+        ln -s $2/dev /data/models/$1/s5/data/wav/
+        echo "Linked dev"
+        cd /data/models/
+     fi
     else
       echo "Linking data from /data/models/$1/wav/ to the data dir of the model"
       ln -s /data/models/$1/wav/train /data/models/$1/s5/data/wav/
